--- a/.github/workflows/build-k3s.yaml
+++ b/.github/workflows/build-k3s.yaml
@@ -21,7 +21,7 @@ permissions:
 
 jobs:
   build:
-    name: Build K3s (${{ inputs.os }} on ${{ inputs.arch }})
+    name: Build K3s
     runs-on: ${{ inputs.arch }}
     timeout-minutes: 20
     env:
@@ -34,12 +34,12 @@ jobs:
       
     - name: Build K3s binary
       run: |
-        DOCKER_BUILDKIT=1 SKIP_IMAGE=1 SKIP_AIRGAP=1 SKIP_VALIDATE=1 GOCOVER=1 GOOS=${{ env.GOOS }} make
+        make local-binary
         sha256sum dist/artifacts/k3s${{ env.BIN_EXT }} | sed 's|dist/artifacts/||' > dist/artifacts/k3s${{ env.BIN_EXT }}.sha256sum
     
     - name: Build K3s image
       if: inputs.upload-image == true && inputs.os == 'linux'
-      run: make package-image
+      run: ./scripts/package-image
       
     - name: "Save K3s image"
       if: inputs.upload-image == true && inputs.os == 'linux'