--- a/.github/workflows/integration.yaml
+++ b/.github/workflows/integration.yaml
@@ -29,9 +29,13 @@ env:
   
 jobs:
   build:
+    permissions:
+      contents: read
+      packages: read
     uses: ./.github/workflows/build-k3s.yaml
     with:
       os: linux
+      cache: read
   build-windows:
     uses: ./.github/workflows/build-k3s.yaml
     with: