--- a/.github/workflows/integration.yaml
+++ b/.github/workflows/integration.yaml
@@ -60,7 +60,7 @@ jobs:
     - name: "Download k3s binary"
       uses: actions/download-artifact@v4
       with:
-        name: k3s
+        name: k3s-amd64
         path: ./dist/artifacts
     - name: Run Integration Tests
       run: |