--- a/scripts/version.sh
+++ b/scripts/version.sh
@@ -4,23 +4,9 @@ GO=${GO-go}
 ARCH=${ARCH:-$("${GO}" env GOARCH)}
 OS=${OS:-$("${GO}" env GOOS)}
 SUFFIX="-${ARCH}"
-GIT_TAG=$DRONE_TAG
-TREE_STATE=clean
-COMMIT=$DRONE_COMMIT
 
-if [ -d .git ]; then
-    if [ -z "$GIT_TAG" ]; then
-        GIT_TAG=$(git tag -l --contains HEAD | head -n 1)
-    fi
-    if [ -n "$(git status --porcelain --untracked-files=no)" ]; then
-        DIRTY="-dirty"
-        TREE_STATE=dirty
-    fi
-
-    COMMIT=$(git log -n3 --pretty=format:"%H %ae" | grep -v ' drone@localhost$' | cut -f1 -d\  | head -1)
-    if [ -z "${COMMIT}" ]; then
-    COMMIT=$(git rev-parse HEAD || true)
-    fi
+if [ -z "$NO_DAPPER" ]; then
+    ./scripts/git_version.sh
 fi
 
 get-module-version(){