… install.sh.sha256sum