<!-- HTML Comments can be left in place or removed. -->
<!-- Please see our contributing guide at https://github.com/k3s-io/k3s/blob/master/CONTRIBUTING.md for guidance on opening pull requests -->

#### Proposed Changes ####
- Add later CI actions of `package` and `airgap` to the `Dockerfile.local`, moving closer to achieving parity with the Dapper Dockerfile
- Breakup version.sh script into git_version.sh and version.sh. This improves caching around build layers, as .git no longer needs to be added to build targets, all git versioning information is calculated beforehand. 

##### Overhaul of `build_k3s.yaml` GHA workflow:
- Rework workflow inputs: `arch` and `cache`. Initial support for emulated builds.
- Introduce layer caching using GHCR. This save about 1min to the build times and uses GHCR (with unlimited storage) vs the dedicated GHA cache (which we only have 10GB of)
- Use faster Dockerfile.local targets where possible, using native GHA calls to docker

<!-- Describe the big picture of your changes here to communicate to the maintainers why we should accept this pull request. -->

#### Types of Changes ####
CI/Build
<!-- What types of changes does your code introduce to K3s? Bugfix, New Feature, Breaking Change, etc -->

#### Verification ####
- CI Still green
- use new make targets `make local-image` or `make local-binary`

## Performance
Setup is as follows:
- 1 is initial build with no cache (this is the most variable and dependent on download time)
- 2 is build with a single line change in pkg/cli (i.e. code change)
- 3 is build with no line changes (just to see maximum caching usage)

| Target | Stage | Dapper | Old Local | New Local |
| - | - | - | - | - |
| Binary | 1 | 3m 22s | 3m 22s | 3m 36s |
| | 2 | 1m 58s | 45s | 40s |
| | 3 | 1m 21s | 2s | 2s |
| Image | 1 | 3m 18s | N/A | 3m 33s |
| | 2 | 1m 36s | N/A | 1m |
| | 3 | 1m 34s | N/A | 3s |

<!-- How can the changes be verified? Please provide whatever additional information necessary to help verify the proposed changes. -->

#### Testing ####
N/A
<!-- Is this change covered by testing? If not, consider adding a Unit or Integration test. -->
<!-- See https://github.com/k3s-io/k3s/blob/master/tests/TESTING.md for more info -->

#### Linked Issues ####
https://github.com/k3s-io/k3s/issues/12136
<!-- Link any related issues, pull-requests, or commit hashes that are relevant to this pull request. If you are opening a PR without a corresponding issue please consider creating one first, at https://github.com/k3s-io/k3s/issues . A functional example will greatly help QA with verifying/reproducing a bug or testing new features. -->

#### User-Facing Change ####
<!--
Does this PR introduce a user-facing change? If no, just write "NONE" in the release-note block below.
If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note

```

#### Further Comments ####
Future work is required to achieve nice build emulation times, with an eye on building arm32bit and potential riscv64. Need to look at QEMU emulation cost or implement `xx` usage, possibly with a separate builds stage.
<!-- If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc... -->